\name{protein_intake}
\alias{protein_intake}
\title{Protein Intake Recommendation}
\description{Estimates daily protein intake based on bodyweight.}
\usage{
protein_intake(bodyweight)
}
\arguments{
  \item{bodyweight}{Numeric body weight in kg.}
}
\value{
Numeric grams of protein per day.
}
\examples{
protein_intake(70)
}